requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0

